requirements.txt
# Core FastAPI dependencies
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2

# File handling and processing
aiofiles==24.1.0
python-multipart==0.0.20
pdf2image==1.17.0
PyMuPDF==1.26.0
pillow==11.2.0

# Authentication and security
python-jose==3.4.0
python-keycloak==5.3.1
jwcrypto==1.5.6
cryptography==45.0.3

# HTTP client
httpx==0.26.0
requests==2.32.3

# Data processing
pandas==2.2.3
numpy==2.2.6
lxml==5.4.0

# Configuration
python-decouple==3.8
pydantic==2.11.5

# Utilities
puremagic==1.29
